feat: use dutch auctions in raft contract (#72)
* Implement Dutch Auction Library and Manager

- Add DutchAuctionLibrary with core auction logic
- Create DutchAuctionManager for multi-auction handling
- Implement custom errors and events
- Develop comprehensive test suite
- Add deployment script and README

This commit introduces a flexible Dutch Auction system supporting
multiple concurrent auctions with configurable parameters. It includes
both linear and exponential price decay models, a kicker reward system,
and thorough error handling. The implementation is fully tested and
ready for review before potential deployment.

* chore: add python script to generate test prices

* chore: fix package.json

* chore: remove gpt response

* chore: add image to readme

* chore: remove unused import

* chore: modify scripts

* Update Codecov badge to current branch

* chore: refactor

* chore: update tests

* chore: optimize auctionConfig struct packing

* fix: formatting

* feat: use dutch auctions in raft contract

* Update Codecov badge to current branch

* chore: cleanup

* chore: add docs

* chore: remove superkeeper

* chore: remove redundant error

* chore: add auction and payment token validation

* chore: remove lcov for dutch auction

* chore: refactor

* chore: remove unnecessary imports

* chore: formatting

* chore: ranme auction parameters

* chore: add multiple auction test cases

* feat: initial buy and burn implementation

* feat: add buy and burn - in progress

* chore: refactor and fix tests

* chore: reuse library errors and events

* Update Codecov badge to current branch

* feat: create a new `percentage` package inside of the monorepo

* chore: add percentage package imports to dutch auction

* chore: update license files

* chore: unify tests, add decimal tests

* chore: create`AuctionManagerBase`

* chore: reafactor tests, add docs

* chore: add tests for different decimal rewards in arks

* chore: change name from `pyament` to `underlying`

* chore: add docs to buy and burn events

* chore: review comments

* chore: fix coverage

* chore: rename test skipper

---------

+++ packages/core-contracts/test/fleets/FleetCommander.redeemWithInterest.t.sol
@@ -210,7 +210,7 @@ contract RedeemWithInterestTest is
 
     function test_RedeemAfterMultipleInterestAccruals() public {
         // Simulate multiple interest accruals
-        for (uint i = 0; i < 3; i++) {
+        for (uint256 i = 0; i < 3; i++) {
             mockToken.mint(ark2, INTEREST_AMOUNT / 3);
         }
